pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pi-weather-ink"
version = "1.4.0"
description = "Raspberry Pi weather station with Waveshare e-Paper displays (multiple models supported)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Benjamin Burzan" },
]
keywords = ["raspberry-pi", "waveshare", "e-paper", "weather-station", "epaper", "epd"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
    "Topic :: System :: Hardware",
]
dependencies = [
    "python-pirateweather>=1.3",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "waveshare-epaper>=1.4.0; platform_system == 'Linux' and platform_machine != 'x86_64'",
    "spidev>=3.8; platform_system == 'Linux'",
    "gpiozero>=2.0.1; platform_system == 'Linux'",
    "lgpio>=0.2.2.0; platform_system == 'Linux' and python_version < '3.13'",
    "rpi-lgpio>=0.6; platform_system == 'Linux' and platform_machine != 'x86_64' and python_version >= '3.13'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "epaper-emulator @ git+https://github.com/b0x42/E-Paper-Emulator.git",
]

[project.scripts]
weatherstation = "weatherstation:main"

[project.urls]
Homepage = "https://github.com/b0x42/pi-weather-ink"
Repository = "https://github.com/b0x42/pi-weather-ink"
Issues = "https://github.com/b0x42/pi-weather-ink/issues"

[tool.setuptools]
py-modules = ["weatherstation", "display_config"]
packages = ["icons"]

[tool.setuptools.package-data]
icons = ["icons.json", "weathericons.ttf"]
